Use a system service for the auto-save polling
This makes the auto-save checks independent of the user's selected game
speed.
We use the cIGZSystemService::OnIdle event, which fires approximately
every 200 milliseconds. This is actually faster than we need because
the minimum save interval is one minute.

Moved the logging code into a singleton class. Changed the log
timestamp format to use local time with the OS localization settings,
instead of [hh:mm:ss.fff] in UTC.

PostAppInit now returns false if any of the setup steps fail.