Disable auto-save when a modal dialog is active
This can occur if the auto save timer is running when the game has a
modal dialog active.
While SC4 pauses the game when a modal dialog is active, for some
reason the cISC4Simulator class does not think the game is paused.

The fix is to detect if a modal dialog is active using the cIGZWinMgr
IsModal method.