Disable auto-save when the game doesn't have focus
According to a comment in the plugin's support thread, SC4 has known
data corruption issues if a save is performed while the app doesn't
have focus (e.g. if it is minimized).

The plugin will also unsubscribe from the OnIdle event when the app
loses focus, there is no point in having that callback fire to check
the elapsed time when saving is disabled until the app regains focus.