Unsubscribe from the OnIdle callback when paused
This prevents the game from wasting CPU on the auto save timer checks.
We never save a city when the game is paused.

Note that this only affects the case where the user has opted-in to the
plugin counting the time the game spends paused towards the next auto
save point.
If the timer is stopped when the game is paused, it already
unsubscribes from the OnIdle callback.